added tape summary output in Brainalyze

+++ src/main/groovy/net/zomis/brainf/model/BrainfuckMemory.groovy
@@ -36,7 +36,7 @@ class BrainfuckMemory {
         }
     }
 
-    public byte[] getMemoryArray(int fromIndex, int length) {
+    public int[] getMemoryArray(int fromIndex, int length) {
         return Arrays.copyOfRange(memory, fromIndex, fromIndex + length);
     }
